+++ CHANGELOG.md
@@ -1,1 +1,7 @@
 # Project Style Checker Change Log
+
+## October 2018
+
+* Extracts `SpyProjectBuilder` to its own project,[spy-project-factory](https://github.com/gmullerb/spy-project-factory), in order to be used by other projects.
+* Upgrades base-style-config to version 1.0.7.
+  * Does some code changes to follow new version.